COMPLITED TASK!!!!

+++ Services/TelstarLogisticsLibraries/CodeBase/Libraries/DbClient/Entitites/PackageContentType.cs
@@ -9,8 +9,8 @@ namespace DbClient.Entitites
     public class PackageContentType : BaseModel
     {
         public string Name { get; set; } = "";
-        public float FeeAmount { get; set; }
+        public double FeeAmount { get; set; }
         public bool IsLegal { get; set; }
-        public float PercentialPriceIncrease { get; set; }
+        public double PercentialPriceIncrease { get; set; }
     }
 }
